make read_string() deal with strings that come in pieces
Older protocol versions sent a ~300k string at start up for some reason.
This doesn't nearly fit in one packet, so we need to keep reading until
we've read all of the string.  While we're there we should really drop
the terminating null byte from the string before returning it.